pyproject.toml
[tool.poetry]
name = "mts"
version = "0.1.0"
description = ""
authors = ["Afanasyev Maxim", "Pushkareva Elizaveta", "Bratchikov Sergey", "Ivanov Arseny"]
packages = [{include = "mts", from = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
light-the-torch = "^0.7.1"
pandas = "^1.5.3"
numpy = "^1.24.1"
tensorboard = "^2.11.2"
xztrainer = "^0.9.3"
jupyterlab = "^3.6.1"
fastparquet = "^2023.1.0"
pyarrow = "^11.0.0"
plotly = "^5.13.0"
ipywidgets = "^8.0.4"
psycopg2 = "^2.9.5"
transformers = "^4.26.1"
scikit-learn = "^1.2.1"
madgrad = "^1.2"
undetected-chromedriver = "^3.4.6"
selenium = "^4.8.0"
orjson = "^3.8.5"
pyvirtualdisplay = "^3.0"
beautifulsoup4 = "^4.11.2"
lxml = "^4.9.2"
sentencepiece = "^0.1.97"
pybind11 = "^2.10.3"
matplotlib = "^3.7.1"
protobuf = "^3.20.0"
lion-pytorch = {git = "https://github.com/lucidrains/lion-pytorch"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
